+++ changelog.md
@@ -1,3 +1,9 @@
+# v1.0.1
+
+* Fixed button being covered by garage button.
+* Made button slightly bigger like 1%.
+* Updated about.md.
+
 # v1.0.0
 
 * First